+++ CHANGES.md
@@ -1,3 +1,8 @@
+## Version 3.0.0
+
+* Add method `paginate()` to `ReadMixin`.
+* Change interface for methods `read()` from `ReadMixin`, `delete()` from `DeleteMixin`.
+
 ## Version 2.1.0
 
 * `PaginateMixin` removed, this functional moved to `ReadMixin`.